Fix Enroll and List percentage

+++ src/biometry/devices/android.cpp
@@ -115,7 +115,7 @@ private:
         {
             if (((androidEnrollOperation*)context)->totalrem == 0)
                 ((androidEnrollOperation*)context)->totalrem = remaining + 1;
-            float raw_value = 1 - (remaining / ((androidEnrollOperation*)context)->totalrem);
+            float raw_value = 1 - ((float)remaining / ((androidEnrollOperation*)context)->totalrem);
             ((androidEnrollOperation*)context)->mobserver->on_progress(biometry::Progress{biometry::Percent::from_raw_value(raw_value), biometry::Dictionary{}});
         } else {
             ((androidEnrollOperation*)context)->mobserver->on_progress(biometry::Progress{biometry::Percent::from_raw_value(1), biometry::Dictionary{}});
@@ -303,7 +303,7 @@ private:
                 ((androidListOperation*)context)->totalrem = remaining + 1;
                 ((androidListOperation*)context)->result.clear();
             }
-            float raw_value = 1 - (remaining / ((androidListOperation*)context)->totalrem);
+            float raw_value = 1 - ((float)remaining / ((androidListOperation*)context)->totalrem);
             ((androidListOperation*)context)->mobserver->on_progress(biometry::Progress{biometry::Percent::from_raw_value(raw_value), biometry::Dictionary{}});
             ((androidListOperation*)context)->result.push_back(fingerId);
         } else {
